divide classes for more organization
